backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
bcrypt
passlib[bcrypt]
python-jose[cryptography]

